gentx-880d6777bc195888cb43a960222c9ec157fdc8a2.json
Thank you for the opportunity to support your mission and participate in the development of the project. Our team believes that this is exactly the time when the world needs a modern financial tool for the Muslim community. We are a team aimed at maintaining and developing blockchain nodes of promising and important projects for the world. 
We will be glad to participate and help in the project.